+++ src/tekton/addtrigger.ts
@@ -22,6 +22,7 @@ import { Progress } from '../util/progress';
 import { cli } from '../cli';
 import { TknVersion, version } from '../util/tknversion';
 import { NewPvc } from './createpvc';
+import { getExposeURl } from '../util/exposeurl';
 
 export const TriggerTemplateModel = {
   apiGroup: 'triggers.tekton.dev',
@@ -101,9 +102,8 @@ export async function k8sCreate(trigger: TriggerTemplateKind | EventListenerKind
     return false;
   }
   if (trigger.kind === RouteModel.kind && !result.error) {
-    const routeInfo = await cli.execute(Command.getRoute(trigger.metadata.name));
-    const routeHost = JSON.parse(routeInfo.stdout).spec.host;
-    vscode.window.showInformationMessage(`Trigger successfully created. Expose URL: http://${routeHost}`);
+    const url = await getExposeURl(trigger.metadata.name);
+    vscode.window.showInformationMessage(`Trigger successfully created. Expose URL: ${url}`);
   }
   await fs.unlink(fsPath);
   return true;
