+++ src/tekton/triggertemplate.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode';
 import { TektonItem } from './tektonitem';
 import { TektonNode, Command, tkn } from '../tkn';
 import { CliCommand } from '../cli';
-import * as _ from 'lodash';
+import { getExposeURl } from '../util/exposeurl';
 
 export class TriggerTemplate extends TektonItem {
 
@@ -26,14 +26,16 @@ export class TriggerTemplate extends TektonItem {
     for (const eventListener of listEventListener) {
       for (const triggers of eventListener.spec.triggers) {
         if (triggers?.template?.name === trigger.getName()) {
-          await TriggerTemplate.getExposeURl(eventListener.status.configuration.generatedName);
+          const url = await getExposeURl(eventListener.status.configuration.generatedName);
+          vscode.env.clipboard.writeText(url);
           vscode.window.showInformationMessage('Expose URl successfully copied');
           return;
         } else if (triggers?.triggerRef) {
           const triggerData = await tkn.execute(Command.getTrigger(triggers.triggerRef));
           const triggerName = JSON.parse(triggerData.stdout).spec.template.name;
           if (triggerName === trigger.getName()) {
-            await TriggerTemplate.getExposeURl(eventListener.status.configuration.generatedName);
+            const url = await getExposeURl(eventListener.status.configuration.generatedName);
+            vscode.env.clipboard.writeText(url);
             vscode.window.showInformationMessage('Expose URl successfully copied');
             return;
           }
@@ -41,16 +43,5 @@ export class TriggerTemplate extends TektonItem {
       }
     }
     vscode.window.showInformationMessage('Expose URl not available');
-  }
-
-  static async getExposeURl(name: string): Promise<void> {
-    const result = await tkn.execute(Command.getRoute(name));
-    const route = JSON.parse(result.stdout);
-    const scheme = _.get(route, 'spec.tls.termination') ? 'https' : 'http';
-    let url = `${scheme}://${route.spec.host}`;
-    if (route.spec?.path) {
-      url += route.spec.path;
-    }
-    vscode.env.clipboard.writeText(url);
   }
 }
